requirements.txt
python-telegram-bot==13.12
openai==0.27.8
ffmpeg-python==0.2.0
whisper==1.0
